conda: new module with 'CondaWrapper' and supporting code extracted from 'pipeliner'.
